making stuff pretty
