+++ src/components/Dashboard/ManageUsersPage/AddUserPage.scss
@@ -4,10 +4,4 @@
   h1 {
     font-size: 4.4rem;
   }
-  label {
-    font-size: 1.5rem;
-  }
-  #email {
-    display: block;
-  } 
 }
