+++ src/components/Dashboard/ManageUsersPage/EditUser.scss
@@ -1,0 +1,6 @@
+.EditUser__name {
+  display: flex;
+  min-width: 200px;
+  max-width: 620px;
+  flex-wrap: wrap;
+}
