+++ src/components/Dashboard/ManageUsersPage/EditUserControl.js
@@ -15,7 +15,7 @@ export default function EditUserControl({
       className={classNames('EditUser__control', className)}
     >
       <div className="EditUser__control-meta">
-        <div className="EditUser__control-label">{`${label}${isRequired && '*'}`}</div>
+        <div className="EditUser__control-label">{`${label}${isRequired ? '*' : ''}`}</div>
       </div>
       <div className="EditUser__control-children">
         {children}
