+++ src/components/Dashboard/ManageUsersPage/EditUserControl.scss
@@ -1,7 +1,7 @@
 .EditUser__control {
   flex-basis: auto;
   min-width: 200px;
-  max-width: 400px;
+  max-width: 800px;
   flex-grow: 1;
   flex-shrink: 0;
   margin: 1rem 1rem;
@@ -19,7 +19,6 @@
 .EditUser__control-required {
   flex: 1;
   text-align: right;
-  font-size: 1.3rem;
   font-weight: 400;
   font-style: italic;
   align-self: center;
@@ -30,4 +29,14 @@
   width: 100%;
   font-size: 1.6rem;
   padding: 1rem;
+  max-width: 600px;
+}
+
+.EditUser__control-radio-button {
+  width: 100%;
+  font-size: 1.4rem;
+  padding: .2rem;
+  input {
+    margin-right: .8rem;
+  }
 }
